[Workflow] Move the dump command to the component
