+++ .github/CODEOWNERS
@@ -55,9 +55,6 @@
 # Workflow
 /src/Symfony/Bridge/Twig/Extension/WorkflowExtension.php @lyrixx
 /src/Symfony/Bridge/Twig/Tests/Extension/WorkflowExtensionTest.php @lyrixx
-/src/Symfony/Bundle/FrameworkBundle/Command/WorkflowDumpCommand.php @lyrixx
-/src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/ValidateWorkflowsPass.php @lyrixx
-/src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/WorkflowGuardListenerPass.php @lyrixx
 /src/Symfony/Component/Workflow/ @lyrixx
 # Yaml
 /src/Symfony/Component/Yaml/ @xabbuh
