+++ src/Symfony/Component/Workflow/composer.json
@@ -26,6 +26,7 @@
     "require-dev": {
         "psr/log": "^1|^2|^3",
         "symfony/config": "^6.4|^7.0|^8.0",
+        "symfony/console": "^6.4|^7.0|^8.0",
         "symfony/dependency-injection": "^6.4|^7.0|^8.0",
         "symfony/error-handler": "^6.4|^7.0|^8.0",
         "symfony/event-dispatcher": "^6.4|^7.0|^8.0",
